1. Fix some context. 2. Move generate random to library and add some example code.
