+++ src/session.h
@@ -29,7 +29,7 @@ extern "C" {
 #define SESSION_CODE_NOT_IMPLEMENTED		HTTP_NOT_IMPLEMENTED
 #define SESSION_CODE_SERVICE_UNAVAILABLE	HTTP_SERVICE_UNAVAILABLE
 
-#define SESSION_MAX_ID 4294967295U
+#define SESSION_MAX_ID RAND_MAX
 #define SESSION_MIN_ID 1
 
 struct model_chain {
